+++ code/game/objects/structures/stool_bed_chair_nest/wheelchair.dm
@@ -1,178 +1,105 @@
 /obj/structure/bed/chair/wheelchair
 	name = "wheelchair"
-	desc = "You sit in this. Either by will or force."
+	desc = "Now we're getting somewhere."
 	icon_state = "wheelchair"
-	anchored = 0
-	buckle_movable = 1
+	anchored = FALSE
+	buckle_movable = TRUE
+	movement_handlers = list(/datum/movement_handler/deny_multiz, /datum/movement_handler/delay = list(5), /datum/movement_handler/move_relay_self)
+	movable_flags = MOVABLE_FLAG_NONDENSE_COLLISION
 
-	var/driving = 0
-	var/mob/living/pulling = null
+	var/item_form_type = /obj/item/wheelchair_kit
+	var/driving = FALSE
 	var/bloodiness
 
-	slowdown = 0
+/obj/structure/bed/chair/wheelchair/Initialize()
+	. = ..()
 
-/obj/structure/bed/chair/wheelchair/update_icon()
-	cut_overlays()
-	add_overlay(image(icon = 'icons/obj/furniture.dmi', icon_state = "w_overlay", layer = FLY_LAYER))
+	if(!item_form_type)
+		verbs -= .verb/collapse
+
+/obj/structure/bed/chair/wheelchair/on_update_icon()
+	return
 
 /obj/structure/bed/chair/wheelchair/set_dir()
-	. = ..()
+	..()
+	overlays.Cut()
+	var/image/O = image(icon = 'icons/obj/furniture.dmi', icon_state = "w_overlay", dir = src.dir)
+	O.layer = ABOVE_HUMAN_LAYER
+	overlays += O
 	if(buckled_mob)
 		buckled_mob.set_dir(dir)
 
-/obj/structure/bed/chair/wheelchair/attackby(obj/item/W as obj, mob/user as mob)
-	if(W.iswrench() || istype(W,/obj/item/stack) || W.iswirecutter())
+/obj/structure/bed/chair/wheelchair/attackby(obj/item/W, mob/user)
+	if(isWrench(W) || istype(W,/obj/item/stack) || isWirecutter(W))
 		return
 	..()
 
 /obj/structure/bed/chair/wheelchair/relaymove(mob/user, direction)
 	// Redundant check?
 	if(user.stat || user.stunned || user.weakened || user.paralysis || user.lying || user.restrained())
-		if(user==pulling)
-			pulling = null
-			user.pulledby = null
-			to_chat(user, "<span class='warning'>You lost your grip!</span>")
-		return
-	if(buckled_mob && pulling && user == buckled_mob)
-		if(pulling.stat || pulling.stunned || pulling.weakened || pulling.paralysis || pulling.lying || pulling.restrained())
-			pulling.pulledby = null
-			pulling = null
-	if(user.pulling && (user == pulling))
-		pulling = null
-		user.pulledby = null
 		return
 	if(propelled)
 		return
-	if(pulling && (get_dist(src, pulling) > 1))
-		pulling = null
-		user.pulledby = null
-		if(user==pulling)
-			return
-	if(pulling && (get_dir(src.loc, pulling.loc) == direction))
-		to_chat(user, "<span class='warning'>You cannot go there.</span>")
-		return
-	if(pulling && buckled_mob && (buckled_mob == user))
-		to_chat(user, "<span class='warning'>You cannot drive while being pushed.</span>")
-		return
-
 	// Let's roll
-	driving = 1
-	var/turf/T = null
+	driving = TRUE
 	//--1---Move occupant---1--//
 	if(buckled_mob)
 		buckled_mob.buckled = null
 		step(buckled_mob, direction)
 		buckled_mob.buckled = src
-	//--2----Move driver----2--//
-	if(pulling)
-		T = pulling.loc
-		if(get_dist(src, pulling) >= 1)
-			step(pulling, get_dir(pulling.loc, src.loc))
-	//--3--Move wheelchair--3--//
+	//--2--Move wheelchair--2--//
 	step(src, direction)
 	if(buckled_mob) // Make sure it stays beneath the occupant
 		Move(buckled_mob.loc)
 	set_dir(direction)
-	if(pulling) // Driver
-		if(pulling.loc == src.loc) // We moved onto the wheelchair? Revert!
-			pulling.forceMove(T)
-		else
-			spawn(0)
-			if(get_dist(src, pulling) > 1) // We are too far away? Losing control.
-				pulling = null
-				user.pulledby = null
-			pulling.set_dir(get_dir(pulling, src)) // When everything is right, face the wheelchair
 	if(bloodiness)
 		create_track()
-	driving = 0
+	driving = FALSE
 
 /obj/structure/bed/chair/wheelchair/Move()
 	. = ..()
-	playsound(src, 'sound/effects/roll.ogg', 75, 1)
 	if(buckled_mob)
 		var/mob/living/occupant = buckled_mob
 		if(!driving)
-			occupant.buckled = null
-			occupant.Move(src.loc)
-			occupant.buckled = src
 			if (occupant && (src.loc != occupant.loc))
 				if (propelled)
 					for (var/mob/O in src.loc)
 						if (O != occupant)
-							Collide(O)
+							Bump(O)
 				else
 					unbuckle_mob()
-			if (pulling && (get_dist(src, pulling) > 1))
-				pulling.pulledby = null
-				to_chat(pulling, "<span class='warning'>You lost your grip!</span>")
-				pulling = null
 		else
 			if (occupant && (src.loc != occupant.loc))
 				src.forceMove(occupant.loc) // Failsafe to make sure the wheelchair stays beneath the occupant after driving
 
-/obj/structure/bed/chair/wheelchair/attack_hand(mob/living/user as mob)
-	if (pulling)
-		MouseDrop(usr)
-	else
-		user_unbuckle_mob(user)
-	return
+/obj/structure/bed/chair/wheelchair/attack_hand(mob/living/user)
+	user_unbuckle_mob(user)
 
-/obj/structure/bed/chair/wheelchair/CtrlClick(var/mob/user)
-	if(in_range(src, user))
-		if(!ishuman(user))	return
-		if(user == buckled_mob)
-			to_chat(user, "<span class='warning'>You realize you are unable to push the wheelchair you sit in.</span>")
-			return
-		if(!pulling)
-			pulling = user
-			user.pulledby = src
-			if(user.pulling)
-				user.stop_pulling()
-			user.set_dir(get_dir(user, src))
-			to_chat(user, "You grip \the [name]'s handles.")
-		else
-			to_chat(usr, "You let go of \the [name]'s handles.")
-			pulling.pulledby = null
-			pulling = null
-		return
+/obj/structure/bed/chair/wheelchair/Bump(atom/A)
+	..()
+	if(!buckled_mob)	return
 
-/obj/structure/bed/chair/wheelchair/Collide(atom/A)
-	. = ..()
-	if(!buckled_mob)
-		return
-
-	if(propelled || (pulling && (pulling.a_intent == I_HURT)))
+	if(propelled)
 		var/mob/living/occupant = unbuckle_mob()
-
-		if (pulling && (pulling.a_intent == I_HURT))
-			occupant.throw_at(A, 3, 3, pulling)
-		else if (propelled)
-			occupant.throw_at(A, 3, propelled)
+		occupant.throw_at(A, 3, 3)
 
 		var/def_zone = ran_zone()
-		var/blocked = occupant.run_armor_check(def_zone, "melee")
-		occupant.throw_at(A, 3, propelled)
+		var/blocked = 100 * occupant.get_blocked_ratio(def_zone, BRUTE, damage = 10)
+		occupant.throw_at(A, 3, 3)
 		occupant.apply_effect(6, STUN, blocked)
 		occupant.apply_effect(6, WEAKEN, blocked)
 		occupant.apply_effect(6, STUTTER, blocked)
-		occupant.apply_damage(10, BRUTE, def_zone, blocked)
-		playsound(src.loc, "punch", 50, 1, -1)
+		occupant.apply_damage(10, BRUTE, def_zone)
+		playsound(src.loc, 'sound/weapons/punch1.ogg', 50, 1, -1)
 		if(istype(A, /mob/living))
 			var/mob/living/victim = A
 			def_zone = ran_zone()
-			blocked = victim.run_armor_check(def_zone, "melee")
+			blocked = 100 * victim.get_blocked_ratio(def_zone, BRUTE, damage = 10)
 			victim.apply_effect(6, STUN, blocked)
 			victim.apply_effect(6, WEAKEN, blocked)
 			victim.apply_effect(6, STUTTER, blocked)
-			victim.apply_damage(10, BRUTE, def_zone, blocked)
-		if(pulling)
-			occupant.visible_message("<span class='danger'>[pulling] has thrusted \the [name] into \the [A], throwing \the [occupant] out of it!</span>")
-
-			pulling.attack_log += "\[[time_stamp()]\]<span class='warning'> Crashed [occupant.name]'s ([occupant.ckey]) [name] into \a [A]</span>"
-			occupant.attack_log += "\[[time_stamp()]\]<font color='orange'> Thrusted into \a [A] by [pulling.name] ([pulling.ckey]) with \the [name]</font>"
-			msg_admin_attack("[pulling.name] ([pulling.ckey]) has thrusted [occupant.name]'s ([occupant.ckey]) [name] into \a [A] (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[pulling.x];Y=[pulling.y];Z=[pulling.z]'>JMP</a>)",ckey=key_name(pulling),ckey_target=key_name(occupant))
-		else
-			occupant.visible_message("<span class='danger'>[occupant] crashed into \the [A]!</span>")
+			victim.apply_damage(10, BRUTE, def_zone)
+		occupant.visible_message(SPAN_DANGER("\The [occupant] crashed into \the [A]!"))
 
 /obj/structure/bed/chair/wheelchair/proc/create_track()
 	var/obj/effect/decal/cleanable/blood/tracks/B = new(loc)
@@ -188,40 +115,56 @@
 		B.set_dir(newdir)
 	bloodiness--
 
-/obj/structure/bed/chair/wheelchair/buckle_mob(mob/M as mob, mob/user as mob)
-	if(M == pulling)
-		pulling = null
-		usr.pulledby = null
-	..()
+/proc/equip_wheelchair(mob/living/carbon/human/H) //Proc for spawning in a wheelchair if a new character has no legs. Used in new_player.dm
+	var/obj/structure/bed/chair/wheelchair/W = new(get_turf(H))
+	if(isturf(H.loc))
+		W.buckle_mob(H)
 
-/obj/item/wheelchair
-	name = "wheelchair"
-	desc = "A folded wheelchair that can be carried around."
-	icon = 'icons/obj/furniture.dmi'
-	icon_state = "wheelchair_folded"
-	item_state = "wheelchair"
-	w_class = ITEMSIZE_HUGE // Can't be put in backpacks. Oh well.
+/obj/structure/bed/chair/wheelchair/verb/collapse()
+	set name = "Collapse Wheelchair"
+	set category = "Object"
+	set src in oview(1)
 
-/obj/item/wheelchair/attack_self(mob/user)
-	var/obj/structure/bed/chair/wheelchair/R = new /obj/structure/bed/chair/wheelchair(user.loc)
-	R.add_fingerprint(user)
-	R.name = src.name
-	R.desc = src.desc
-	R.color = src.color
-	qdel(src)
+	if(!item_form_type)
+		return
 
-/obj/structure/bed/chair/wheelchair/MouseDrop(over_object, src_location, over_location)
-	..()
-	if((over_object == usr && (in_range(src, usr) || usr.contents.Find(src))))
-		if(!ishuman(usr))
-			return FALSE
-		if(buckled_mob)
-			return FALSE
-	visible_message(SPAN_NOTICE("[usr] collapses [src]."))
-	var/obj/item/wheelchair/R = new(get_turf(src))
-	R.name = src.name
-	R.desc = src.desc
-	R.color = src.color
-	usr.put_in_hands(R)
-	qdel(src)
+	if(!CanPhysicallyInteract(usr))
+		return
 
+	if(!ishuman(usr))
+		return
+
+	if(usr.incapacitated())
+		return
+
+	if(buckled_mob)
+		to_chat(usr, SPAN_WARNING("You can't collapse \the [src.name] while it is still in use."))
+		return
+
+	usr.visible_message("<b>[usr]</b> starts to collapse \the [src.name].")
+	if(do_after(usr, 4 SECONDS, src))
+		var/obj/item/wheelchair_kit/K = new item_form_type(get_turf(src))
+		visible_message(SPAN_NOTICE("<b>[usr]</b> collapses \the [src.name]."))
+		K.add_fingerprint(usr)
+		qdel(src)
+
+/obj/item/wheelchair_kit
+	name = "compressed wheelchair kit"
+	desc = "Collapsed parts, prepared to immediately spring into the shape of a wheelchair."
+	icon = 'icons/obj/items/wheelchairkit.dmi'
+	icon_state = "wheelchair-item"
+	item_state = "rbed"
+	w_class = ITEM_SIZE_LARGE
+
+	var/structure_form_type = /obj/structure/bed/chair/wheelchair
+
+/obj/item/wheelchair_kit/attack_self(mob/user)
+	if(!structure_form_type)
+		return
+
+	user.visible_message("<b>[user]</b> starts to lay out \the [src.name].")
+	if(do_after(user, 4 SECONDS, src))
+		var/obj/structure/bed/chair/wheelchair/W = new structure_form_type(get_turf(user))
+		user.visible_message(SPAN_NOTICE("<b>[user]</b> lays out \the [W.name]."))
+		W.add_fingerprint(user)
+		qdel(src)
